feat: create specific error
