+++ src/Collection/Collection.router.ts
@@ -19,7 +19,7 @@ import {
   UnpublishedItemError,
   InconsistentItemError,
 } from '../Item/Item.errors'
-import { Item, ThirdPartyItemAttributes, ItemApprovalData } from '../Item'
+import { Item, ItemApprovalData } from '../Item'
 import { isCommitteeMember } from '../Committee'
 import {
   buildCollectionForumPost,
@@ -30,7 +30,6 @@ import {
 } from '../Forum'
 import { sendDataToWarehouse } from '../warehouse'
 import { Cheque } from '../SlotUsageCheque'
-import { ThirdPartyService } from '../ThirdParty/ThirdParty.service'
 import { hasTPCollectionURN, isTPCollection } from '../utils/urn'
 import { Collection } from './Collection.model'
 import { CollectionService } from './Collection.service'
@@ -44,6 +43,7 @@ import { hasPublicAccess } from './access'
 import { toFullCollection } from './utils'
 import {
   AlreadyPublishedCollectionError,
+  InsufficientSlotsError,
   LockedCollectionError,
   NonExistentCollectionError,
   UnauthorizedCollectionEditError,
@@ -55,7 +55,6 @@ const validator = getValidator()
 
 export class CollectionRouter extends Router {
   public service = new CollectionService()
-  public tpService = new ThirdPartyService()
 
   private modelAuthorizationCheck = (
     _: OwnableModel,
@@ -314,23 +313,10 @@ export class CollectionRouter extends Router {
 
       if (isTPCollection(dbCollection)) {
         const itemIds = server.extractFromReq<string[]>(req, 'itemIds')
-        const availableSlots = await this.tpService.getThirdPartyAvailableSlots(
-          dbCollection.third_party_id
-        )
-        if (itemIds.length > availableSlots) {
-          throw new HTTPError(
-            'The amount of items to publish exceeds the available slots',
-            null,
-            STATUS_CODES.badRequest
-          )
-        }
-        const dbItems = (await Item.findByIds(
-          itemIds
-        )) as ThirdPartyItemAttributes[]
 
         result = await this.service.publishTPCollection(
+          itemIds,
           dbCollection,
-          dbItems,
           eth_address,
           server.extractFromReq<Cheque>(req, 'cheque')
         )
@@ -387,6 +373,12 @@ export class CollectionRouter extends Router {
           STATUS_CODES.conflict
         )
       } else if (error instanceof AlreadyPublishedCollectionError) {
+        throw new HTTPError(
+          error.message,
+          { id: error.id },
+          STATUS_CODES.conflict
+        )
+      } else if (error instanceof InsufficientSlotsError) {
         throw new HTTPError(
           error.message,
           { id: error.id },
